agents/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langgraph==0.0.60
anthropic==0.7.8
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.110.0
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.5.2
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
